Improved support for third party SB branches and added 64bit support for thetvdb integration
Fixes #49

+++ app/src/main/java/com/rastating/droidbeard/ShowSearch.java
@@ -75,7 +75,7 @@ public class ShowSearch extends Activity implements ApiResponseListener<TvDBResu
             @Override
             public void onItemClick(AdapterView<?> adapterView, View view, int i, long l) {
                 TextView idView = (TextView) view.findViewById(R.id.id);
-                int id = Integer.valueOf(idView.getText().toString());
+                long id = Long.valueOf(idView.getText().toString());
                 showProgressDialog("Adding Show", "Please wait...");
 
                 AddShowTask task = new AddShowTask(ShowSearch.this);
